+++ Modes.h
@@ -61,7 +61,7 @@ typedef enum eMode : byte
 // Overload the ControlType++ operator
 inline Mode& operator++(Mode& eDOW, int)  // <--- note -- must be a reference
 {
-  const int i = static_cast<int>(eDOW) + 1;
+  const byte i = static_cast<byte>(eDOW) + 1;
   eDOW = static_cast<Mode>((i) % EXT_MODE_COUNT);
   return eDOW;
 }
